Add files via upload
These .ino files are loaded onto an Arduino Mega 2560 or Arduino Mini to get two WS2812B LED strips to light up the StormBreaker Axe as Blue ombre or pulsing (blinking lightning wasn't rulebound) then turn orange when a button is pushed, or Red ombre and turn orange when a button is pushed.  Arduino pins 2 and 4 are used for LED output and Button input.